Support QingStor API
This commit refactors class HttpConnection, make it to be a more
general HTTP client wrapper, then implements QSConnection (which
manages connection to QingStor) with the refactored HttpConnection
as the base class.

A new module named qingstor is added to support all the QingStor API.
It hides complexity and details of HTTP API with an object-oriented
API. Also add related unit tests for this module incidentally.

Last but not least, QingCloud SDK python 1.0.0 is released!